+++ runtime/devnet/src/pallets/mod.rs
@@ -6,3 +6,6 @@ pub use grandpa::*;
 
 pub mod timestamp;
 pub use timestamp::*;
+
+pub mod balances;
+pub use balances::*;
